[frontend] Move json helper to own package
By now we have quite a few helper that take received JSON responses from
the API and convert them to neat little domain objects. This patch moves
those classes to their own package.
