+++ frontend/src/store/modules/comparisonGraphStore.ts
@@ -7,7 +7,7 @@ import {
 } from '@/store/types'
 import Vue from 'vue'
 import axios from 'axios'
-import { comparisonDatapointFromJson } from '@/util/GraphJsonHelper'
+import { comparisonDatapointFromJson } from '@/util/json/GraphJsonHelper'
 import { vxm } from '@/store'
 import router from '@/router'
 import { PermanentLinkOptions } from '@/store/modules/detailGraphStore'
