+++ frontend/src/util/json/GraphJsonHelper.ts
